pyproject.toml
[tool.poetry]
name = "books-store"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
dj-database-url = "^2.1.0"
djangorestframework = "^3.14.0"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.61.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
